Crate linalg

A linear algebra library with BLAS and LAPACK acceleration

Using linalg

Add these dependencies to your Cargo.toml

[dependencies.linalg]
features = ["macros"]  # Optional, enables the mat! syntax extension
git = "https://github.com/japaric/linalg.rs"

All the functionality is available via the prelude module, glob import it:

// Optional, enables the mat! syntax extension
#![plugin(linalg_macros)]

extern crate linalg;

// Imports extension traits and main `struct`s like `Mat`
use linalg::prelude::*;

Quick reference

For NumPy/Octave users

Matrix initialization

// Octave
A = [1, 2, 3; 4, 5, 6; 7, 8, 9];

// Rust
let A = mat![1, 2, 3; 4, 5, 6; 7, 8, 9];

Indexing

// Python
A[1, 2] = 5
x = A[3, 4]

// Rust
A[(1, 2)] = 5;
let x = A[(3, 4)];  // or `&A[(3, 4)]` or `&mut A[(3, 4)]`

Slicing

// Python
second_row = A[1, :]
third_column = A[:, 2]
submat = A[:3, 2:]  // or A[0:3, 2:4]

// Rust
let second_row = A.row(1);        // or A.slice((1, ..))
let third_column = A.col(2);      // or A.slice((.., 2))
let submat = A.slice((..3, 2..))  // or A.slice((0..3, 2..4))

// NOTE All the operations have a mutable variant, just add a `_mut` suffix.
// Example: `let second_row = A.row_mut(1)`;

Augmented assignment

Increase all the elements of the second row by 1.

// Python
A[1, :] += 1;

// Rust
A.row_mut(1).add_assign(1)

Subtract sub-matrices

// Python
A[1:3, 2:4] -= B[:2, 1:3]

// Rust
A.slice_mut((1..3, 2..4)).sub_assign(B.slice((..2, 1..3)));

Index assignment

Set all the elements of the second column to 0

// Python
A[:, 1] = 0;

// Rust
A.col_mut(1).set(0);

Copy sub-matrices

// Python
A[1:3, 1:3] = B[2:4, 3:5]

// Rust
A.slice_mut((1..3, 1..3)).set(B.slice((2..4, 3..5)));

Matrix multiplication

// Python
D = A.dot(B).dot(C)

// Rust
let D = A * B * C;

Transpose

// Python
A = B.T

// Rust
let A = B.t();

Matrix inverse

// Python
B = numpy.linalg.inv(A)

// Rust
let B = A.inv();

Overview of the API

There are two types of structures provided by this crate: "owned structures", and "views". The former own the data, and are in charge of freeing memory when droped, the latter are borrows of the former, and provide limited access to the data.

The owned structures: Mat, ColVec and RowVec provide constructors via static methods.

Most of the functionality (other than overloaded operators) is provided as methods via extension traits. All the public extension traits are grouped in the traits module.

There are several conversions between views and from views to owned structures available via the From/Into traits.

Notes about operators

Keep in mind that all operators (unary/binary) take their operands by value.

Both forms of multiplication: matrix multiplication and scaling are lazy. If you want eager evaluation, you can use the eval() method, but this allocates memory, and should be avoided whenever possible.

// No op
let C = A * B;

// Performs the matrix multiplication
let C = (A * B).eval();

When multiplying matrices, you never want to give up ownership. Always multiply views.

// A and B are owned
let C = A * B;  //~ error

let C = &A * &B;  // OK

Addition and subtraction are eager, and one of the operands must provide a buffer to store the result, in most cases this means that one of the operators needs to be moved into the operation.

// Both A and B own their data

let C = A + &B;  // OK, result will be stored in A's buffer

let C = A + B;
//~^ error: the result could be stored in A's buffer, but B would be unnecessarily dropped

let C = &A + &B;  //~ error: nowhere to store the result

When you only have views and need to perform an addition/subtraction, you'll have to allocate memory. For performance, avoid allocating in loops.

// theta, X and y are views

// Bad: allocation in a loop
loop {
    let y = ColVec::from(y);  // Deep copy
    let e = y - theta * X;

    ..

    if condition(&e) { break }
}

// Instead allocate a buffer outside the loop, and re-use it in the loop
let mut z = ColVec::zeros(y.nrows());

loop {
    // z = y
    z.set(y);

    // z = y - theta * X
    z.sub_assign(theta * X);

    ..

    if condition(&z) { break }
}

Transposing a matrix is "free", no allocations, copies or operations are performed . Do note that the t() method takes the caller by value.

The inv() method computes the inverse of an owned (square) matrix and takes ownership of the caller. The caller's buffer will be re-used to store the inverse.

Modules

prelude
The "prelude", a collection of the most used traits and structs.

strided
Strided slices

submat
Sub-matrix views

traits
Extension traits

transposed
Transposed matrices

Structs

Chain
Lazy matrix chain multiplication

Col
Immutable view into the column of a matrix

ColMut
Mutable "view" into the column of a matrix

ColVec
Owned column vector

Cols
Iterator over the columns of an immutable matrix

ColsMut
Iterator over the columns of a mutable matrix

Diag
An immutable view into the diagonal of a matrix

DiagMut
A mutable "view" into the diagonal of a matrix

HStripes
An immutable iterator over a matrix in horizontal stripes

HStripesMut
A "mutable" iterator over a matrix in horizontal stripes

Mat
An owned matrix

Product
Lazy matrix product

Row
Immutable view into the row of a matrix

RowMut
Mutable "view" into the row of a matrix

RowVec
An owned row vector

Rows
Iterator over the rows of an immutable matrix

RowsMut
Iterator over the rows of a mutable matrix

Scaled
A lazily scaled matrix

SubMat
Immutable sub-matrix view

SubMatMut
Mutable sub-matrix "view"

Transposed
A lazily transposed matrix

VStripes
An immutable iterator over a matrix in vertical stripes

VStripesMut
A "mutable" iterator over a matrix in vertical stripes