Struct linalg::ColVec

pub struct ColVec<T>(_);

An owned column vector

Methods

impl<T> ColVec<T>

fn new(data: Box<[T]>) -> ColVec<T>

Creates a column vector from an existing vector

Examples

assert_eq!(ColVec::new(box [0i, 1, 2]), mat![0i; 1; 2])

fn from_fn<F>(length: uint, f: F) -> ColVec<T> where F: FnMut(uint) -> T

Creates a column vector and initializes each element to f(index)

Memory: O(length)

Time: O(length)

Examples

assert_eq!(ColVec::from_fn(3, |i| i), mat![0; 1; 2])

fn sample<D, R>(length: uint, distribution: &D, rng: &mut R) -> ColVec<T> where D: IndependentSample<T>, R: Rng

Creates a column vector and fills it by sampling a random distribution

Memory: O(length)

Time: O(length)

fn len(&self) -> uint

Returns the length of the column

impl<T> ColVec<T> where T: Clone

fn from_elem(length: uint, value: T) -> ColVec<T>

Constructs a column vector with copies of a value

Memory: O(length)

Time: O(length)

Examples

assert_eq!(ColVec::from_elem(3, 2), mat![2i; 2; 2])

impl<T> ColVec<T> where T: Rand

fn rand<R>(length: uint, rng: &mut R) -> ColVec<T> where R: Rng

Constructs a randomly initialized column vector

Memory: O(length)

Time: O(length)