Struct linalg::Scaled

pub struct Scaled<T, M>(_, _);

A lazily scaled matrix

Methods

impl<T, M> Scaled<T, M> where M: MatrixCols, T: Clone

fn cols(&self) -> Scaled<T, Cols<M>>

Returns an iterator that yields immutable views into the columns of the matrix

impl<T, M> Scaled<T, M> where M: MatrixRows, T: Clone

fn rows(&self) -> Scaled<T, Rows<M>>

Returns an iterator that yields immutable views into each row of the matrix