Struct linalg::Transposed

pub struct Transposed<M>(_);

A lazily transposed matrix

Methods

impl<M> Transposed<M>

fn col(&self, i: u32) -> Col<M> where M: MatrixRow

Returns an immutable "view" into the ith column of the matrix

fn col_mut(&mut self, i: u32) -> ColMut<M> where M: MatrixRowMut

Returns a mutable "view" into the ith column of the matrix

fn cols(&self) -> Cols<M> where M: MatrixRows

Returns an iterator that yields immutable view into the columns of the matrix

fn cols_mut(&mut self) -> ColsMut<M> where M: MatrixRowsMut

Returns an iterator that yields mutable "views" into the columns of the matrix

fn diag(&self, i: i32) -> Diag<M> where M: MatrixDiag

Returns an immutable view into the ith diagonal of the matrix

fn diag_mut(&mut self, i: i32) -> DiagMut<M> where M: MatrixDiagMut

Returns a mutable "view" into the ith diagonal of the matrix

fn hsplit_at(&self, i: u32) -> (Transposed<SubMat<M>>, Transposed<SubMat<M>>) where M: VSplit

Splits a matrix horizontally at the ith row in two immutable pieces

fn hsplit_at_mut(&mut self, i: u32) -> (Transposed<SubMatMut<M>>, Transposed<SubMatMut<M>>) where M: VSplitMut

Splits a matrix horizontally at the ith row in two mutable pieces

fn hstripes(&self, size: u32) -> HStripes<M> where M: MatrixVStripes

Returns an immutable iterator that yields horizontal stripes of size rows

fn hstripes_mut(&mut self, size: u32) -> HStripesMut<M> where M: MatrixVStripesMut

Returns an "mutable" iterator that yields horizontal stripes of size rows

fn iter<'a>(&'a self) -> M where M: Iter<'a>

Returns an iterator that yields immutable references to the elements of the matrix

NOTE For optimization reasons the iteration order is left unspecified, so don't rely on it

fn iter_mut<'a>(&'a mut self) -> M where M: IterMut<'a>

Returns an iterator that yields mutable references to the elements of the matrix

NOTE For optimization reasons the iteration order is left unspecified, so don't rely on it

fn row(&self, i: u32) -> Row<M> where M: MatrixCol

Returns an immutable "view" into the ith row of the matrix

fn row_mut(&mut self, i: u32) -> RowMut<M> where M: MatrixColMut

Returns a mutable "view" into the ith row of the matrix

fn rows(&self) -> Rows<M> where M: MatrixCols

Returns an iterator that yields immutable views into the rows of a matrix

fn rows_mut(&mut self) -> RowsMut<M> where M: MatrixColsMut

Returns an iterator that yields mutable "views" into the rows a matrix

fn vsplit_at(&self, i: u32) -> (Transposed<SubMat<M>>, Transposed<SubMat<M>>) where M: HSplit

Splits a matrix vertically at the ith column in two immutable pieces

fn vsplit_at_mut(&mut self, i: u32) -> (Transposed<SubMatMut<M>>, Transposed<SubMatMut<M>>) where M: HSplitMut

Splits a matrix vertically at the ith column in two mutable pieces

fn vstripes(&self, size: u32) -> VStripes<M> where M: MatrixHStripes

Returns an immutable iterator that yields vertical stripes of size columns

fn vstripes_mut(&mut self, size: u32) -> VStripesMut<M> where M: MatrixHStripesMut

Returns a "mutable" iterator that yields vertical stripes of size columns