Module linalg::traits

Extension traits

WARNING! At this point in time, these traits are not meant to be used for generic programming, and will remain unstable. What's guaranteed to be (somewhat) stable is the functionality, i.e. the methods, provided by them.

Traits

Eval
Force evaluation of lazy operations

HSplit
"Immutable" horizontal splitting

HSplitMut
"Mutable" horizontal splitting

Iter
"Immutable iteration" over a matrix

IterMut
"Mutable iteration" over a matrix

Matrix
The basic idea of a matrix: A rectangular array arranged in rows and columns

MatrixCol
Immutable view into the column of a matrix

MatrixColMut
Mutable access to the column of a matrix

MatrixCols
Immutable column-by-column iteration

MatrixColsMut
Mutable column-by-column iteration

MatrixDiag
Immutable view into the diagonal of a matrix

MatrixDiagMut
Mutable access to the diagonal of a matrix

MatrixHStripes
"Immutable iteration" over a matrix in horizontal stripes

MatrixHStripesMut
"Mutable iteration" over a matrix in horizontal stripes

MatrixInverse
Matrix inverse

MatrixRow
Immutable view into the row of a matrix

MatrixRowMut
Mutable access to the row of a matrix

MatrixRows
Immutable row-by-row iteration

MatrixRowsMut
Mutable row-by-row iteration

MatrixVStripes
"Immutable iteration" over a matrix in vertical stripes

MatrixVStripesMut
"Mutable iteration" over a matrix in vertical stripes

Norm
Frobenius norm

Set
Alternative to IndexSet (which doesn't exist)

Slice
"Immutable slicing"

SliceMut
"Mutable slicing"

Transpose
The transpose operator

VSplit
"Immutable" vertical splitting

VSplitMut
"Mutable" vertical splitting