Trait linalg::traits::Slice

pub trait Slice<'a, I, S> {
    fn slice(&'a self, start: I, end: I) -> Result<S>;
    fn slice_from(&'a self, start: I) -> Result<S>;
    fn slice_to(&'a self, end: I) -> Result<S>;
}

A more flexible slicing trait

Note Sadly this doesn't have operator sugar. You won't be able to use the slicing operator [] with this library until Rust gets HKT.

Required Methods

fn slice(&'a self, start: I, end: I) -> Result<S>

Returns an immutable view into a fraction of the collection that spans start : end

fn slice_from(&'a self, start: I) -> Result<S>

Convenience method for slice(start, end_of_collection)

fn slice_to(&'a self, end: I) -> Result<S>

Convenience method for slice(start_of_collection, end)