Crate linalg

An experimental linear algebra library with BLAS (*) acceleration

(*): This library is continuously tested against OpenBLAS and netlib/reference BLAS

Cargo

Cargo.toml

[dependencies.linalg]
git = "https://github.com/japaric/linalg.rs"

[dependencies.linalg_macros]
git = "https://github.com/japaric/linalg.rs"

Crate file

extern crate linalg;
#[phase(plugin)]
extern crate linalg_macros;

use linalg::prelude::*;

Design goals

Make memory allocations explicit

Minimize the creation of temporaries

Accelerate (BLAS/SIMD) everything

Conventions

All operations are O(1) in time and memory unless otherwise noted

Matrices are laid in memory using column-major order

Element-wise iteration over [sub]matrices is done in the fastest way possible, there's no guarantee of the iteration order

Modules

blas
BLAS acceleration

prelude
A collection of the most used structs and traits, meant to be glob imported

strided
Strided slices

traits
Traits

Structs

Col
Immutable view into the column of a matrix

ColVec
An owned column vector

Cols
Iterator over the columns of an immutable matrix

Diag
Immutable view into the diagonal of a matrix

Items
Immutable sub-matrix iterator

Mat
Owned matrix

MutCol
Mutable view into the column of a matrix

MutCols
Iterator over the columns of a mutable matrix

MutDiag
Immutable view into the diagonal of a matrix

MutItems
Mutable sub-matrix iterator

MutRow
Mutable view into the row of a matrix

MutRows
Iterator over the rows of a mutable matrix

MutView
Mutable sub-matrix view

Row
Immutable view into the row of a matrix

RowVec
An owned row vector

Rows
Iterator over the rows of an immutable matrix

Scaled
A lazily scaled matrix

Trans
View into the transpose of a matrix

View
Immutable sub-matrix view

Enums

Error
Errors

Type Definitions

Result
The result of a matrix operation